Move /next into main entrypoint
